Strip out D2 concepts from old loadout drawer
