+++ config/webpack.js
@@ -411,8 +411,6 @@ module.exports = (env) => {
         '$featureFlags.abilityCooldowns': JSON.stringify(false),
         // Item feed sidebar
         '$featureFlags.itemFeed': JSON.stringify(true),
-        // Loadout edit drawer v2
-        '$featureFlags.loadoutDrawerV2': JSON.stringify(true),
       }),
 
       new LodashModuleReplacementPlugin({
