+++ src/app/loadout-drawer/LoadoutDrawerContents.tsx
@@ -1,35 +1,27 @@
 import { t } from 'app/i18next-t';
 import { storesSelector } from 'app/inventory/selectors';
-import { SocketOverrides, SocketOverridesForItems } from 'app/inventory/store/override-sockets';
+import { SocketOverrides } from 'app/inventory/store/override-sockets';
 import { getCurrentStore, getStore } from 'app/inventory/stores-helpers';
-import { pickSubclass } from 'app/loadout/item-utils';
 import { D1BucketHashes } from 'app/search/d1-known-values';
 import { itemCanBeInLoadout } from 'app/utils/item-utils';
 import { infoLog } from 'app/utils/log';
-import { getSocketsByCategoryHash } from 'app/utils/socket-utils';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
-import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
+import { BucketHashes } from 'data/d2/generated-enums';
 import produce from 'immer';
 import _ from 'lodash';
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 import type { InventoryBucket, InventoryBuckets } from '../inventory/inventory-buckets';
-import { DimItem, PluggableInventoryItemDefinition } from '../inventory/item-types';
+import { DimItem } from '../inventory/item-types';
 import { DimStore } from '../inventory/store-types';
 import { showItemPicker } from '../item-picker/item-picker';
-import { addIcon, AppIcon, faTshirt } from '../shell/icons';
+import { addIcon, AppIcon } from '../shell/icons';
 import { Loadout, LoadoutItem } from './loadout-types';
-import {
-  createSocketOverridesFromEquipped,
-  createSubclassDefaultSocketOverrides,
-  extractArmorModHashes,
-  fromEquippedTypes,
-} from './loadout-utils';
+import { fromEquippedTypes } from './loadout-utils';
 import LoadoutDrawerBucket from './LoadoutDrawerBucket';
-import SavedMods from './SavedMods';
-import { Subclass } from './subclass-drawer/Subclass';
 
 const loadoutTypes: (BucketHashes | D1BucketHashes)[] = [
+  BucketHashes.Subclass,
   BucketHashes.KineticWeapons,
   BucketHashes.EnergyWeapons,
   BucketHashes.PowerWeapons,
@@ -54,33 +46,23 @@ const loadoutTypes: (BucketHashes | D1BucketHashes)[] = [
 export default function LoadoutDrawerContents({
   storeId,
   loadout,
-  savedMods,
   buckets,
   items,
   equip,
   remove,
   add,
   onUpdateLoadout,
-  onOpenModPicker,
   onShowItemPicker,
-  onOpenFashionDrawer,
-  removeModByHash,
-  onApplySocketOverrides,
 }: {
   storeId?: string;
   loadout: Loadout;
-  savedMods: PluggableInventoryItemDefinition[];
   buckets: InventoryBuckets;
   items: DimItem[];
   equip(item: DimItem, e: React.MouseEvent): void;
   remove(item: DimItem, e: React.MouseEvent): void;
   add(params: { item: DimItem; equip?: boolean; socketOverrides?: SocketOverrides }): void;
   onUpdateLoadout(loadout: Loadout): void;
-  onOpenModPicker(): void;
   onShowItemPicker(shown: boolean): void;
-  onOpenFashionDrawer(): void;
-  removeModByHash(itemHash: number): void;
-  onApplySocketOverrides(item: DimItem, socketOverrides: SocketOverrides): void;
 }) {
   const stores = useSelector(storesSelector);
   const itemsByBucket = _.groupBy(items, (i) => i.bucket.hash);
@@ -92,14 +74,9 @@ export default function LoadoutDrawerContents({
         stores.find((s) => s.classType === loadout.classType)) ||
       getCurrentStore(stores)!;
 
-  function doFillLoadoutFromEquipped(e: React.MouseEvent) {
-    e.preventDefault();
+  const doFillLoadoutFromEquipped = () =>
     fillLoadoutFromEquipped(loadout, itemsByBucket, dimStore, onUpdateLoadout);
-  }
-  function doFillLoadOutFromUnequipped(e: React.MouseEvent) {
-    e.preventDefault();
-    fillLoadoutFromUnequipped(loadout, dimStore, add);
-  }
+  const doFillLoadOutFromUnequipped = () => fillLoadoutFromUnequipped(loadout, dimStore, add);
 
   const availableTypes = _.compact(loadoutTypes.map((h) => buckets.byHash[h]));
 
@@ -107,48 +84,28 @@ export default function LoadoutDrawerContents({
     availableTypes,
     (bucket) => bucket.hash && itemsByBucket[bucket.hash]?.length
   );
-  const subclassBucket = buckets.byHash[BucketHashes.Subclass];
 
   const showFillFromEquipped = typesWithoutItems.some((b) => fromEquippedTypes.includes(b.hash));
-
-  const { subclassItems, subclassSocketOverrides } = useMemo(() => {
-    const subclassSocketOverrides: SocketOverridesForItems = {};
-    const subclassItems = itemsByBucket[BucketHashes.Subclass] ?? [];
-
-    for (const item of loadout.items) {
-      if (subclassItems.some((subclass) => subclass.id === item.id)) {
-        subclassSocketOverrides[item.id] = item.socketOverrides || {};
-      }
-    }
-    return { subclassSocketOverrides, subclassItems };
-  }, [itemsByBucket, loadout.items]);
-
-  const showSubclassButton =
-    !loadout ||
-    loadout.classType === DestinyClass.Unknown ||
-    !subclassItems.length ||
-    subclassItems.every((i) => i.classType !== loadout.classType);
 
   return (
     <>
       <div className="loadout-add-types">
         {showFillFromEquipped && (
-          <a className="dim-button loadout-add" onClick={doFillLoadoutFromEquipped}>
-            <AppIcon icon={addIcon} /> {t('Loadouts.AddEquippedItems')}
-          </a>
-        )}
-        <a className="dim-button loadout-add" onClick={doFillLoadOutFromUnequipped}>
-          <AppIcon icon={addIcon} /> {t('Loadouts.AddUnequippedItems')}
-        </a>
-        {showSubclassButton && (
-          <a
-            key={subclassBucket.hash}
-            onClick={() => pickLoadoutSubclass(loadout, subclassItems, add, onShowItemPicker)}
+          <button
+            type="button"
             className="dim-button loadout-add"
+            onClick={doFillLoadoutFromEquipped}
           >
-            <AppIcon icon={addIcon} /> {subclassBucket.name}
-          </a>
+            <AppIcon icon={addIcon} /> {t('Loadouts.AddEquippedItems')}
+          </button>
         )}
+        <button
+          type="button"
+          className="dim-button loadout-add"
+          onClick={doFillLoadOutFromUnequipped}
+        >
+          <AppIcon icon={addIcon} /> {t('Loadouts.AddUnequippedItems')}
+        </button>
         {typesWithoutItems.length > 0 &&
           typesWithoutItems.map((bucket) => (
             <a
@@ -159,51 +116,25 @@ export default function LoadoutDrawerContents({
               <AppIcon icon={addIcon} /> {bucket.name}
             </a>
           ))}
-        <a onClick={() => onOpenModPicker()} className="dim-button loadout-add">
-          <AppIcon icon={addIcon} /> {t('Loadouts.ArmorMods')}
-        </a>
-        {loadout.classType !== DestinyClass.Unknown && (
-          <a onClick={() => onOpenFashionDrawer()} className="dim-button loadout-add">
-            <AppIcon icon={faTshirt} /> {t('Loadouts.Fashion')}
-          </a>
-        )}
       </div>
       <div className="loadout-added-items">
-        {typesWithItems.map((bucket) =>
-          bucket.hash === BucketHashes.Subclass ? null : (
-            <LoadoutDrawerBucket
-              key={bucket.hash}
-              bucket={bucket}
-              loadoutItems={loadout.items}
-              items={itemsByBucket[bucket.hash] || []}
-              pickLoadoutItem={(bucket) => pickLoadoutItem(loadout, bucket, add, onShowItemPicker)}
-              equip={equip}
-              remove={remove}
-            />
-          )
-        )}
-      </div>
-      {subclassItems.length > 0 &&
-        subclassItems.map((subclass) => (
-          <Subclass
-            key={subclass.hash}
-            subclass={subclass}
-            socketOverrides={subclassSocketOverrides[subclass.id]}
+        {typesWithItems.map((bucket) => (
+          <LoadoutDrawerBucket
+            key={bucket.hash}
+            bucket={bucket}
+            loadoutItems={loadout.items}
+            items={itemsByBucket[bucket.hash] || []}
+            pickLoadoutItem={(bucket) => pickLoadoutItem(loadout, bucket, add, onShowItemPicker)}
             equip={equip}
             remove={remove}
-            onApplySocketOverrides={onApplySocketOverrides}
           />
         ))}
-      <SavedMods
-        savedMods={savedMods}
-        onOpenModPicker={onOpenModPicker}
-        removeModByHash={removeModByHash}
-      />
+      </div>
     </>
   );
 }
 
-export async function pickLoadoutItem(
+async function pickLoadoutItem(
   loadout: Loadout,
   bucket: InventoryBucket,
   add: (params: { item: DimItem }) => void,
@@ -239,47 +170,9 @@ export async function pickLoadoutItem(
   }
 }
 
-export async function pickLoadoutSubclass(
+function fillLoadoutFromEquipped(
   loadout: Loadout,
-  savedSubclasses: DimItem[],
-  add: (params: { item: DimItem; socketOverrides?: SocketOverrides }) => void,
-  onShowItemPicker: (shown: boolean) => void
-) {
-  const loadoutClassType = loadout?.classType;
-  const loadoutHasItem = (item: DimItem) =>
-    loadout?.items.some((i) => i.id === item.id && i.hash === item.hash);
-
-  const loadoutHasSubclassForClass = (item: DimItem) =>
-    savedSubclasses.some(
-      (s) => item.bucket.hash === BucketHashes.Subclass && s.classType === item.classType
-    );
-
-  const subclassItemFilter = (item: DimItem) =>
-    item.bucket.hash === BucketHashes.Subclass &&
-    (!loadout ||
-      loadout.classType === DestinyClass.Unknown ||
-      item.classType === loadoutClassType) &&
-    itemCanBeInLoadout(item) &&
-    !loadoutHasSubclassForClass(item) &&
-    !loadoutHasItem(item);
-
-  onShowItemPicker(true);
-  const item = await pickSubclass(subclassItemFilter);
-  if (item) {
-    let socketOverrides: SocketOverrides | undefined;
-    if (item.bucket.hash === BucketHashes.Subclass) {
-      socketOverrides = createSubclassDefaultSocketOverrides(item);
-    }
-
-    add({ item, socketOverrides });
-  }
-  onShowItemPicker(false);
-}
-
-/** Replace the loadout's subclass with the currently equipped subclass */
-export function setLoadoutSubclassFromEquipped(
-  loadout: Loadout,
-  existingSubclass: DimItem | undefined,
+  itemsByBucket: { [bucketId: string]: DimItem[] },
   dimStore: DimStore,
   onUpdateLoadout: (loadout: Loadout) => void
 ) {
@@ -287,53 +180,9 @@ export function setLoadoutSubclassFromEquipped(
     return;
   }
 
-  const newSubclass = dimStore.items.find(
-    (item) =>
-      item.equipped && itemCanBeInLoadout(item) && item.bucket.hash === BucketHashes.Subclass
-  );
-
-  if (!newSubclass) {
-    return;
-  }
-
-  const newLoadoutItem: LoadoutItem = {
-    id: newSubclass.id,
-    hash: newSubclass.hash,
-    equipped: true,
-    amount: 1,
-    socketOverrides: createSocketOverridesFromEquipped(newSubclass),
-  };
-
-  const newLoadout = {
-    ...loadout,
-    items: [...loadout.items.filter((i) => existingSubclass?.hash !== i.hash), newLoadoutItem],
-  };
-
-  onUpdateLoadout(newLoadout);
-}
-
-export function fillLoadoutFromEquipped(
-  loadout: Loadout,
-  itemsByBucket: { [bucketId: string]: DimItem[] },
-  dimStore: DimStore,
-  onUpdateLoadout: (loadout: Loadout) => void,
-  // This is a bit dangerous as it is only used from the new loadout edit drawer and
-  // has special handling that would break the old loadout drawer
-  category?: string
-) {
-  if (!loadout) {
-    return;
-  }
-
   const newEquippedItems = dimStore.items.filter(
     (item) =>
-      item.equipped &&
-      itemCanBeInLoadout(item) &&
-      (category
-        ? category === 'General'
-          ? item.bucket.hash !== BucketHashes.Subclass && item.bucket.sort === category
-          : item.bucket.sort === category
-        : fromEquippedTypes.includes(item.bucket.hash))
+      item.equipped && itemCanBeInLoadout(item) && fromEquippedTypes.includes(item.bucket.hash)
   );
 
   const hasEquippedInBucket = (bucket: InventoryBucket) =>
@@ -345,7 +194,6 @@ export function fillLoadoutFromEquipped(
     );
 
   const newLoadout = produce(loadout, (draftLoadout) => {
-    const mods: number[] = [];
     for (const item of newEquippedItems) {
       if (!hasEquippedInBucket(item.bucket)) {
         const loadoutItem: LoadoutItem = {
@@ -354,47 +202,17 @@ export function fillLoadoutFromEquipped(
           equipped: true,
           amount: 1,
         };
-        if (item.bucket.hash === BucketHashes.Subclass) {
-          loadoutItem.socketOverrides = createSocketOverridesFromEquipped(item);
-        }
         draftLoadout.items.push(loadoutItem);
-        mods.push(...extractArmorModHashes(item));
       } else {
         infoLog('loadout', 'Skipping', item, { itemsByBucket, bucketId: item.bucket.hash });
       }
-    }
-    if (mods.length && (loadout.parameters?.mods ?? []).length === 0) {
-      draftLoadout.parameters = {
-        ...draftLoadout.parameters,
-        mods,
-      };
-    }
-    // Save "fashion" mods for equipped items
-    const modsByBucket = {};
-    for (const item of newEquippedItems.filter((i) => i.bucket.inArmor)) {
-      const plugs = item.sockets
-        ? _.compact(
-            getSocketsByCategoryHash(item.sockets, SocketCategoryHashes.ArmorCosmetics).map(
-              (s) => s.plugged?.plugDef.hash
-            )
-          )
-        : [];
-      if (plugs.length) {
-        modsByBucket[item.bucket.hash] = plugs;
-      }
-    }
-    if (!_.isEmpty(modsByBucket)) {
-      draftLoadout.parameters = {
-        ...draftLoadout.parameters,
-        modsByBucket,
-      };
     }
   });
 
   onUpdateLoadout(newLoadout);
 }
 
-export async function fillLoadoutFromUnequipped(
+async function fillLoadoutFromUnequipped(
   loadout: Loadout,
   dimStore: DimStore,
   add: (params: { item: DimItem; equip?: boolean }) => void,
@@ -416,7 +234,7 @@ export async function fillLoadoutFromUnequipped(
  * filter for items that are in a character's "pockets" but not equipped,
  * and can be added to a loadout
  */
-export function getUnequippedItemsForLoadout(dimStore: DimStore, category?: string) {
+function getUnequippedItemsForLoadout(dimStore: DimStore, category?: string) {
   return dimStore.items.filter(
     (item) =>
       !item.location.inPostmaster &&
