--- src/app/loadout-drawer/SavedModCategory.m.scss
@@ -1,33 +1,0 @@
-.category {
-  display: block;
-  margin-right: 16px;
-}
-
-.categoryNameContainer {
-  display: flex;
-}
-
-.categoryName {
-  flex-grow: 1;
-  font-size: 10px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  text-transform: uppercase;
-  white-space: nowrap;
-  width: 0;
-}
-
-.mods {
-  align-items: center;
-  display: flex;
-  flex-flow: row wrap;
-  width: fit-content;
-
-  > * {
-    margin-right: 4px;
-
-    &:last-child {
-      margin-right: 0;
-    }
-  }
-}
