--- src/app/loadout-drawer/SavedModCategory.m.scss.d.ts
@@ -1,10 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'category': string;
-  'categoryName': string;
-  'categoryNameContainer': string;
-  'mods': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
