--- src/app/loadout-drawer/SavedMods.m.scss
@@ -1,14 +1,0 @@
-.container {
-  margin: 8px 0;
-}
-
-.title {
-  font-size: 13px;
-  margin-right: 4px 0;
-  text-transform: uppercase;
-}
-
-.categories {
-  display: flex;
-  flex-flow: row wrap;
-}
