--- src/app/loadout-drawer/SavedMods.m.scss.d.ts
@@ -1,9 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'categories': string;
-  'container': string;
-  'title': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
