+++ src/app/loadout-drawer/loadout-drawer-reducer.ts
@@ -24,12 +24,6 @@ export interface State {
   storeId?: string;
   showClass: boolean;
   isNew: boolean;
-  modPicker: {
-    show: boolean;
-    /** An initial query to be passed to the mod picker, this will filter the mods shown. */
-    query?: string;
-  };
-  showFashionDrawer: boolean;
 }
 
 export type Action =
@@ -64,10 +58,7 @@ export type Action =
   | { type: 'equipItem'; item: DimItem; items: DimItem[] }
   | { type: 'updateMods'; mods: number[] }
   | { type: 'changeClearMods'; enabled: boolean }
-  | { type: 'removeMod'; hash: number }
-  | { type: 'openModPicker'; query?: string }
-  | { type: 'closeModPicker' }
-  | { type: 'toggleFashionDrawer'; show: boolean };
+  | { type: 'removeMod'; hash: number };
 
 /**
  * All state for this component is managed through this reducer and the Actions above.
@@ -79,10 +70,6 @@ export function stateReducer(state: State, action: Action): State {
         showClass: true,
         isNew: false,
         loadout: undefined,
-        modPicker: {
-          show: false,
-        },
-        showFashionDrawer: false,
       };
 
     case 'editLoadout': {
@@ -250,181 +237,171 @@ export function stateReducer(state: State, action: Action): State {
       }
       return state;
     }
-
-    case 'openModPicker': {
-      const { query } = action;
-      return { ...state, modPicker: { show: true, query } };
-    }
-
-    case 'closeModPicker': {
-      return { ...state, modPicker: { show: false } };
-    }
-
-    case 'toggleFashionDrawer':
-      return { ...state, showFashionDrawer: action.show };
   }
-}
 
-/**
- * Produce a new loadout that adds a new item to the given loadout.
- */
-function addItem(
-  loadout: Readonly<Loadout>,
-  item: DimItem,
-  shift: boolean,
-  items: DimItem[],
-  equip?: boolean,
-  socketOverrides?: SocketOverrides
-): Loadout {
-  const loadoutItem: LoadoutItem = {
-    id: item.id,
-    hash: item.hash,
-    amount: Math.min(item.amount, shift ? 5 : 1),
-    equipped: false,
-  };
+  /**
+   * Produce a new loadout that adds a new item to the given loadout.
+   */
+  function addItem(
+    loadout: Readonly<Loadout>,
+    item: DimItem,
+    shift: boolean,
+    items: DimItem[],
+    equip?: boolean,
+    socketOverrides?: SocketOverrides
+  ): Loadout {
+    const loadoutItem: LoadoutItem = {
+      id: item.id,
+      hash: item.hash,
+      amount: Math.min(item.amount, shift ? 5 : 1),
+      equipped: false,
+    };
 
-  // TODO: maybe we should just switch back to storing loadout items in memory by bucket
+    // TODO: maybe we should just switch back to storing loadout items in memory by bucket
 
-  // Other items of the same type (as DimItem)
-  const typeInventory = items.filter((i) => i.bucket.hash === item.bucket.hash);
-  const dupe = loadout.items.find((i) => i.hash === item.hash && i.id === item.id);
-  const maxSlots = item.bucket.capacity;
+    // Other items of the same type (as DimItem)
+    const typeInventory = items.filter((i) => i.bucket.hash === item.bucket.hash);
+    const dupe = loadout.items.find((i) => i.hash === item.hash && i.id === item.id);
+    const maxSlots = item.bucket.capacity;
 
-  return produce(loadout, (draftLoadout) => {
-    const findItem = (item: DimItem) =>
-      draftLoadout.items.find((i) => i.id === item.id && i.hash === item.hash)!;
+    return produce(loadout, (draftLoadout) => {
+      const findItem = (item: DimItem) =>
+        draftLoadout.items.find((i) => i.id === item.id && i.hash === item.hash)!;
 
-    if (!dupe) {
-      if (typeInventory.length < maxSlots) {
-        loadoutItem.equipped =
-          equip !== undefined ? equip : item.equipment && typeInventory.length === 0;
-        if (loadoutItem.equipped) {
-          for (const otherItem of typeInventory) {
-            findItem(otherItem).equipped = false;
+      if (!dupe) {
+        if (typeInventory.length < maxSlots) {
+          loadoutItem.equipped =
+            equip !== undefined ? equip : item.equipment && typeInventory.length === 0;
+          if (loadoutItem.equipped) {
+            for (const otherItem of typeInventory) {
+              findItem(otherItem).equipped = false;
+            }
           }
+
+          // Only allow one subclass to be present per class (to allow for making a loadout that specifies a subclass for each class)
+          if (item.bucket.hash === BucketHashes.Subclass) {
+            const conflictingItem = items.find(
+              (i) => i.bucket.hash === item.bucket.hash && i.classType === item.classType
+            );
+            if (conflictingItem) {
+              draftLoadout.items = draftLoadout.items.filter((i) => i.id !== conflictingItem.id);
+            }
+            loadoutItem.equipped = true;
+          }
+
+          if (socketOverrides) {
+            loadoutItem.socketOverrides = socketOverrides;
+          }
+
+          draftLoadout.items.push(loadoutItem);
+
+          // If adding a new armor item, remove any fashion mods (shader/ornament) that couldn't be slotted
+          if (
+            item.bucket.inArmor &&
+            loadoutItem.equipped &&
+            draftLoadout.parameters?.modsByBucket?.[item.bucket.hash]?.length
+          ) {
+            const cosmeticSockets = getSocketsByCategoryHash(
+              item.sockets,
+              SocketCategoryHashes.ArmorCosmetics
+            );
+            draftLoadout.parameters.modsByBucket[item.bucket.hash] =
+              draftLoadout.parameters.modsByBucket[item.bucket.hash].filter((plugHash) =>
+                cosmeticSockets.some((s) =>
+                  s.plugSet?.plugs.some((p) => p.plugDef.hash === plugHash)
+                )
+              );
+          }
+        } else {
+          showNotification({
+            type: 'warning',
+            title: t('Loadouts.MaxSlots', { slots: maxSlots }),
+          });
         }
+      } else if (item.maxStackSize > 1) {
+        const increment = Math.min(dupe.amount + item.amount, item.maxStackSize) - dupe.amount;
+        dupe.amount += increment;
+      }
+    });
+  }
 
-        // Only allow one subclass to be present per class (to allow for making a loadout that specifies a subclass for each class)
-        if (item.bucket.hash === BucketHashes.Subclass) {
-          const conflictingItem = items.find(
-            (i) => i.bucket.hash === item.bucket.hash && i.classType === item.classType
+  /**
+   * Produce a new Loadout with the given item removed from the original loadout.
+   */
+  function removeItem(
+    loadout: Readonly<Loadout>,
+    item: DimItem,
+    shift: boolean,
+    items: DimItem[]
+  ): Loadout {
+    return produce(loadout, (draftLoadout) => {
+      const loadoutItem = draftLoadout.items.find((i) => i.hash === item.hash && i.id === item.id);
+
+      if (!loadoutItem) {
+        return;
+      }
+
+      const decrement = shift ? 5 : 1;
+      loadoutItem.amount ||= 1;
+      loadoutItem.amount -= decrement;
+      if (loadoutItem.amount <= 0) {
+        draftLoadout.items = draftLoadout.items.filter(
+          (i) => !(i.hash === item.hash && i.id === item.id)
+        );
+      }
+
+      if (loadoutItem.equipped) {
+        const typeInventory = items.filter((i) => i.bucket.hash === item.bucket.hash);
+        const nextInLine =
+          typeInventory.length > 0 &&
+          draftLoadout.items.find(
+            (i) => i.id === typeInventory[0].id && i.hash === typeInventory[0].hash
           );
-          if (conflictingItem) {
-            draftLoadout.items = draftLoadout.items.filter((i) => i.id !== conflictingItem.id);
-          }
+        if (nextInLine) {
+          nextInLine.equipped = true;
+        }
+      }
+    });
+  }
+
+  /**
+   * Produce a new loadout with the given item switched to being equipped (or unequipped if it's already equipped).
+   */
+  function equipItem(loadout: Readonly<Loadout>, item: DimItem, items: DimItem[]) {
+    return produce(loadout, (draftLoadout) => {
+      const findItem = (item: DimItem) =>
+        draftLoadout.items.find((i) => i.id === item.id && i.hash === item.hash)!;
+
+      // Classes are always equipped
+      if (item.bucket.hash === BucketHashes.Subclass) {
+        return;
+      }
+
+      const loadoutItem = findItem(item);
+      if (item.equipment) {
+        if (loadoutItem.equipped) {
+          // It's equipped, mark it unequipped
+          loadoutItem.equipped = false;
+        } else {
+          // It's unequipped - mark all the other items and conflicting exotics unequipped, then mark this equipped
+          items
+            .filter(
+              (i) =>
+                // Others in this slot
+                i.bucket.hash === item.bucket.hash ||
+                // Other exotics
+                (item.equippingLabel && i.equippingLabel === item.equippingLabel)
+            )
+            .map(findItem)
+            .forEach((i) => {
+              i.equipped = false;
+            });
+
           loadoutItem.equipped = true;
         }
-
-        if (socketOverrides) {
-          loadoutItem.socketOverrides = socketOverrides;
-        }
-
-        draftLoadout.items.push(loadoutItem);
-
-        // If adding a new armor item, remove any fashion mods (shader/ornament) that couldn't be slotted
-        if (
-          item.bucket.inArmor &&
-          loadoutItem.equipped &&
-          draftLoadout.parameters?.modsByBucket?.[item.bucket.hash]?.length
-        ) {
-          const cosmeticSockets = getSocketsByCategoryHash(
-            item.sockets,
-            SocketCategoryHashes.ArmorCosmetics
-          );
-          draftLoadout.parameters.modsByBucket[item.bucket.hash] =
-            draftLoadout.parameters.modsByBucket[item.bucket.hash].filter((plugHash) =>
-              cosmeticSockets.some((s) => s.plugSet?.plugs.some((p) => p.plugDef.hash === plugHash))
-            );
-        }
-      } else {
-        showNotification({
-          type: 'warning',
-          title: t('Loadouts.MaxSlots', { slots: maxSlots }),
-        });
       }
-    } else if (item.maxStackSize > 1) {
-      const increment = Math.min(dupe.amount + item.amount, item.maxStackSize) - dupe.amount;
-      dupe.amount += increment;
-    }
-  });
-}
-
-/**
- * Produce a new Loadout with the given item removed from the original loadout.
- */
-function removeItem(
-  loadout: Readonly<Loadout>,
-  item: DimItem,
-  shift: boolean,
-  items: DimItem[]
-): Loadout {
-  return produce(loadout, (draftLoadout) => {
-    const loadoutItem = draftLoadout.items.find((i) => i.hash === item.hash && i.id === item.id);
-
-    if (!loadoutItem) {
-      return;
-    }
-
-    const decrement = shift ? 5 : 1;
-    loadoutItem.amount ||= 1;
-    loadoutItem.amount -= decrement;
-    if (loadoutItem.amount <= 0) {
-      draftLoadout.items = draftLoadout.items.filter(
-        (i) => !(i.hash === item.hash && i.id === item.id)
-      );
-    }
-
-    if (loadoutItem.equipped) {
-      const typeInventory = items.filter((i) => i.bucket.hash === item.bucket.hash);
-      const nextInLine =
-        typeInventory.length > 0 &&
-        draftLoadout.items.find(
-          (i) => i.id === typeInventory[0].id && i.hash === typeInventory[0].hash
-        );
-      if (nextInLine) {
-        nextInLine.equipped = true;
-      }
-    }
-  });
-}
-
-/**
- * Produce a new loadout with the given item switched to being equipped (or unequipped if it's already equipped).
- */
-function equipItem(loadout: Readonly<Loadout>, item: DimItem, items: DimItem[]) {
-  return produce(loadout, (draftLoadout) => {
-    const findItem = (item: DimItem) =>
-      draftLoadout.items.find((i) => i.id === item.id && i.hash === item.hash)!;
-
-    // Classes are always equipped
-    if (item.bucket.hash === BucketHashes.Subclass) {
-      return;
-    }
-
-    const loadoutItem = findItem(item);
-    if (item.equipment) {
-      if (loadoutItem.equipped) {
-        // It's equipped, mark it unequipped
-        loadoutItem.equipped = false;
-      } else {
-        // It's unequipped - mark all the other items and conflicting exotics unequipped, then mark this equipped
-        items
-          .filter(
-            (i) =>
-              // Others in this slot
-              i.bucket.hash === item.bucket.hash ||
-              // Other exotics
-              (item.equippingLabel && i.equippingLabel === item.equippingLabel)
-          )
-          .map(findItem)
-          .forEach((i) => {
-            i.equipped = false;
-          });
-
-        loadoutItem.equipped = true;
-      }
-    }
-  });
+    });
+  }
 }
 
 function applySocketOverrides(
