--- src/app/loadout-drawer/subclass-drawer/ItemSocketsSubclass.m.scss
@@ -1,49 +1,0 @@
-@import '../../variables';
-@import '../../dim-ui/common.m.scss';
-
-.category {
-  margin-top: 8px;
-  margin-right: 16px;
-}
-
-.title {
-  margin-bottom: 4px;
-  font-size: 10px;
-}
-
-.socketCategory {
-  &:not(:last-child) {
-    margin-right: 4px;
-  }
-
-  img {
-    height: var(--item-size);
-    width: var(--item-size);
-  }
-}
-
-.socket {
-  border: 1px solid #ddd;
-}
-
-.super {
-  position: relative;
-  border: 1px solid transparent;
-
-  > img {
-    display: block;
-  }
-}
-
-.superOutline {
-  position: absolute;
-  left: 0;
-  top: 0;
-
-  height: var(--item-size);
-  width: var(--item-size);
-}
-
-.sockets {
-  composes: flexRow;
-}
