--- src/app/loadout-drawer/subclass-drawer/ItemSocketsSubclass.m.scss.d.ts
@@ -1,17 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'category': string;
-  'flexColumn': string;
-  'flexRow': string;
-  'flexWrap': string;
-  'resetButton': string;
-  'socket': string;
-  'socketCategory': string;
-  'sockets': string;
-  'super': string;
-  'superOutline': string;
-  'title': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
