--- src/app/loadout-drawer/subclass-drawer/Subclass.m.scss
@@ -1,24 +1,0 @@
-@import '../../dim-ui/common.m.scss';
-
-.container {
-  composes: flexWrap;
-  align-items: flex-end;
-
-  > * {
-    &:not(:last-child) {
-      margin-right: 16px;
-    }
-  }
-}
-
-.subclass {
-  :global(.item) {
-    height: calc(1.1 * var(--item-size));
-    width: calc(1.1 * var(--item-size));
-
-    img {
-      height: calc(1.1 * var(--item-size));
-      width: calc(1.1 * var(--item-size));
-    }
-  }
-}
