--- src/app/loadout-drawer/subclass-drawer/Subclass.m.scss.d.ts
@@ -1,12 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'container': string;
-  'flexColumn': string;
-  'flexRow': string;
-  'flexWrap': string;
-  'resetButton': string;
-  'subclass': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
