+++ src/app/shell/Destiny.tsx
@@ -256,11 +256,7 @@ export default function Destiny() {
           <Route path="*" element={<Navigate to="inventory" />} />
         </Routes>
       </div>
-      {$featureFlags.loadoutDrawerV2 && account.destinyVersion === 2 ? (
-        <LoadoutDrawer2 />
-      ) : (
-        <LoadoutDrawer />
-      )}
+      {account.destinyVersion === 2 ? <LoadoutDrawer2 /> : <LoadoutDrawer />}
       <Compare />
       <Farming />
       <InfusionFinder />
